Add macro for transitive dependencies
This was really tough to figure out - I used rules_docker as a guide,
but it took quite a while to realize that I would need to copy their
"maybe"-esque logic for conditionally adding go_repository rules.

Without this it seems that the @org_golang_x_tools repository in
_rules_nfpm_go_repositories replaces the transitive @org_golang_x_tools
repository from @io_bazel_rules_go. Due to some implementation-specific
patches in @io_bazel_rules_go this would cause a 'target not found'
error as they had not been applied.

What was particularly odd is that this would only occur if the call to
_rules_nfpm_go_repositories happened inside of
rules_nfpm_go_dependencies, but not if it was called separately in the
WORKSPACE file.

Unfortunately I could not find a way to make gazelle use
maybe_go_repository when generating repository information, but I was
able to trick it by binding it to go_repository.